+++ schema.sql
@@ -1,5 +1,15 @@
-CREATE TABLE IF NOT EXISTS leaderboard (
+CREATE TABLE IF NOT EXISTS tournaments (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    name TEXT UNIQUE NOT NULL,
+    short_name TEXT UNIQUE NOT NULL
+);
+
+CREATE TABLE IF NOT EXISTS tournament_entries (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    tournament_id INTEGER NOT NULL,
     user_id TEXT PRIMARY KEY,
     username TEXT NOT NULL,
-    wave INTEGER NOT NULL
+    waves INTEGER NOT NULL
+    FOREIGN KEY (tournament_id) REFERENCES tournaments(id) ON DELETE CASCADE,
+    UNIQUE (tournament_id, user_id)
 );